requirements.txt
Flask
gradio
huggingface-hub
jieba
torch
torchvision
Werkzeug
openai
langchain
spacy
flask-cors
requests_toolbelt
transformers
openai-whisper
cchardet
websocket-client
azure-cognitiveservices-speech
